Change type of index argument in calculate_dataset_item()

+++ lib/ethash/ethash-internal.hpp
@@ -46,6 +46,6 @@ hash256 calculate_seed(int epoch_number) noexcept;
 
 void build_light_cache(hash512 cache[], int num_items, const hash256& seed) noexcept;
 
-hash1024 calculate_dataset_item(const epoch_context& context, size_t index) noexcept;
+hash1024 calculate_dataset_item(const epoch_context& context, uint32_t index) noexcept;
 
 }  // namespace ethash
